Counter von int zu long konvertiert

+++ Studienarbeit/DataReader/src/main/java/de/dhbw/studienarbeit/data/helper/database/table/DatabaseTable.java
@@ -62,7 +62,7 @@ public abstract class DatabaseTable extends DatabaseConnector
 
 	protected abstract String getTableName();
 
-	public int count() throws IOException
+	public long count() throws IOException
 	{
 		final String countEntries = "SELECT COUNT(*) AS total FROM " + getTableName() + ";";
